pyproject.toml
[tool.poetry]
name = "conflixis-analytics"
version = "0.1.0"
description = "Collaborative data analytics repository"
authors = ["Conflixis Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
numpy = "^2.0.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
jupyter = "^1.1.0"
ipykernel = "^6.29.0"
google-cloud-bigquery = "^3.25.0"
google-cloud-bigquery-storage = "^2.26.0"
pyarrow = "^20.0.0"
db-dtypes = "^1.1.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.5"
pypdf2 = "^3.0.1"
jellyfish = "^1.2.0"
rapidfuzz = "^3.13.0"
snowflake-connector-python = "^3.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^25.1.0"
flake8 = "^7.1.0"
mypy = "^1.13.0"
pre-commit = "^4.2.0"
ruff = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"